Refactored migrations

+++ data/migrations/20191104151245_albums.js
@@ -5,10 +5,10 @@ exports.up = function(knex) {
   return knex.schema.createTable('albums', table =>  {
 
     // album_id
-    table.bigIncrements('album_id').primary();
+    table.increments('album_id').primary();
 
     // user_id
-    table.bigInteger('user_id')
+    table.integer('user_id')
       .unsigned()
       .notNullable()
       .references('user_id')
@@ -17,7 +17,7 @@ exports.up = function(knex) {
       .onDelete('CASCADE');
 
     // cover_id
-    table.bigInteger('cover_id')
+    table.integer('cover_id')
       .unsigned()
       .defaultTo(null);
