+++ data/migrations/20191104174802_comments.js
@@ -5,10 +5,10 @@ exports.up = function(knex) {
   return knex.schema.createTable('comments', table => {
 
     // comment_id
-    table.bigIncrements('comment_id').primary();
+    table.increments('comment_id').primary();
 
     // user_id
-    table.bigInteger('user_id')
+    table.integer('user_id')
       .notNullable()
       .references('user_id')
       .inTable('users')
@@ -16,7 +16,7 @@ exports.up = function(knex) {
       .onDelete('CASCADE');
 
     // media_id
-    table.bigInteger('media_id')
+    table.integer('media_id')
       .defaultTo(null)
       .references('media_id')
       .inTable('media')
@@ -24,7 +24,7 @@ exports.up = function(knex) {
       .onDelete('CASCADE');
 
     // album_id
-    table.bigInteger('album_id')
+    table.integer('album_id')
       .defaultTo(null)
       .references('album_id')
       .inTable('albums')
@@ -32,7 +32,7 @@ exports.up = function(knex) {
       .onDelete('CASCADE');
 
     // parent_id
-    table.bigInteger('parent_id')
+    table.integer('parent_id')
       .defaultTo(null)
       .references('comment_id')
       .inTable('comments')
